Added axis label multiline support in canvas
All \n line breaks in an axis label will now be displayed as such in a canvas axis label. Also, the axis label text can be automatically broken into multiple lines (if wished) if the length of the axis isn't enough. The alignment of and padding between this lines can be choosen.
Added axis options are...
axisLabelAdjustment (true/false): If the axis label should be broken into multiple lines at spaces in case the length of the axis isn't enough for the width of the label. Default is false.
axisLabelLinePadding (number): Padding between multiple axis label lines. Default is 5.
axisLabelAlignment ('left', 'center'/'middle', 'right): Alignment of multiple axis label lines (aligned according to widest line and reading direction). Default is 'center'.

Also: Fix in calculating canvas label size, the 0.72 multiplier was missing (but used when calculating position of line).